CloudSQL_With_Public_IP/requirements.txt
Flask==2.3.2
google-cloud==0.34.0
psycopg2==2.9.9 
cloud-sql-python-connector==1.4.1
SQLAlchemy==2.0.21
pg8000==1.30.2